Move memory tests to crates/levm/tests/
